Add support for passing a URL; restore created_by; add import
Rather than hacking the script for every use, let's make it more
configurable, allowing the caller to pass a URL.

created_by should default to upload.py surely, as it used to?
Make it "osm-bulk-upload/upload.py" to make it clearer which fork it is.

Also import=yes seems common practice, add that.

See:
https://www.openstreetmap.org/changeset/91920753 (by someone else)
https://www.openstreetmap.org/changeset/94937072 (by me)